Refactor Features tab components into presentational and container componenets
