+++ static/js/components/FeaturizeForm/index.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { reduxForm } from 'redux-form';
 import ReactTabs from 'react-tabs';
 
-import * as Validate from '../validate';
 import { FormComponent, Form, TextInput, TextareaInput, SubmitButton,
-         CheckBoxInput, SelectInput } from './Form';
-import Expand from './Expand';
-import { contains } from '../utils';
-import * as Action from '../actions';
+         CheckBoxInput, SelectInput } from '../Form';
+import Expand from '../Expand';
+import { contains } from '../../utils';
+import * as Action from '../../actions';
 
 
 const { Tab, Tabs, TabList, TabPanel } = { ...ReactTabs };
@@ -131,38 +129,4 @@ FeaturizeForm.defaultProps = {
   error: ""
 };
 
-const mapStateToProps = (state, ownProps) => {
-  const featuresList = state.features.featsWithCheckedTags;
-
-  const initialValues = { };
-  featuresList.map((f, idx) => { initialValues[f] = true; return null; });
-
-  const filteredDatasets = state.datasets.filter(dataset =>
-    (dataset.project_id === ownProps.selectedProject.id));
-  const zerothDataset = filteredDatasets[0];
-
-  return {
-    featuresByCategory: state.features.features_by_category,
-    tagList: state.features.tagList,
-    featuresList,
-    featureDescriptions: state.features.descriptions,
-    datasets: filteredDatasets,
-    fields: featuresList.concat(
-      ['datasetID', 'featuresetName', 'customFeatsCode']
-    ),
-    initialValues: { ...initialValues,
-                     datasetID: zerothDataset ? zerothDataset.id.toString() : "",
-                     customFeatsCode: "" }
-  };
-};
-
-const validate = Validate.createValidator({
-  datasetID: [Validate.required],
-  featuresetName: [Validate.required]
-});
-
-export default reduxForm({
-  form: 'featurize',
-  fields: [''],
-  validate
-}, mapStateToProps)(FeaturizeForm);
+export default FeaturizeForm;
