+++ static/js/components/UploadFeaturesForm/index.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { reduxForm } from 'redux-form';
 
 import { FormComponent, Form, TextInput, TextareaInput, SubmitButton,
-         CheckBoxInput, SelectInput, FileInput } from './Form';
-import * as Validate from '../validate';
-import CesiumTooltip from './Tooltip';
+         CheckBoxInput, SelectInput, FileInput } from '../Form';
+import CesiumTooltip from '../Tooltip';
 
 
 const UploadFeaturesForm = props => {
@@ -59,27 +57,4 @@ UploadFeaturesForm.defaultProps = {
   error: ""
 };
 
-const mapStateToProps = (state, ownProps) => {
-  const initialValues = { };
-  const filteredDatasets = state.datasets.filter(dataset =>
-    (dataset.project_id === ownProps.selectedProject.id));
-  return {
-    datasets: filteredDatasets,
-    fields: ['datasetID', 'featuresetName', 'dataFile'],
-    initialValues: { ...initialValues,
-                     datasetID: "No associated dataset" }
-  };
-};
-
-const validate = Validate.createValidator({
-  featuresetName: [Validate.required],
-  dataFile: [Validate.oneFile]
-});
-
-export default reduxForm(
-  {
-    form: 'uploadFeatures',
-    fields: [''],
-    validate
-  }, mapStateToProps
-)(UploadFeaturesForm);
+export default UploadFeaturesForm;
